change number of person in menu to double

+++ src/main/java/com/aws/codestar/projecttemplates/dto/MenuDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class MenuDTO {
     private Long id;
 
-    private int numberOfPeople;
+    private double numberOfPeople;
 
     private String name;
